Move ArrayTableWidget and Model to silx.gui.widgets
